requirements.txt
# Core NLP
vaderSentiment>=3.3.2
bertopic>=0.16.0
sentence-transformers>=2.2.2

# ML & Embeddings
scikit-learn>=1.3.0
umap-learn>=0.5.4
hdbscan>=0.8.33

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Text Processing
nltk>=3.8.1
regex>=2023.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wordcloud>=1.9.0

# Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Utilities
joblib>=1.3.0
